Revert "[sc-116510] only install ghcup"
This reverts commit fdffdd6d8899a3df520eb8d630049828ca42d3b8.

This didn't work because when trying to run `./develop.sh` you'll
get:

```
stack_1                     | ./tools/docker/run-ide.sh: line 5:
stack: command not found
```

This is because stack is no longer installed.

Adding only ghcup and stack isn't an option, because we no longer let
stack manage the toolchain.

The minimum set of things to install for ghcup to manage the
toolchain and for things to work as they were is:

- ghcup
- ghc
- cabal
- stack

We can wait to install haskell-language-server in the
docker-stack-develop container though.